feat(core): improve code
