+++ packages/core/src/__tests__/form.state.spec.ts
@@ -1,9 +1,9 @@
-import { FormState } from '../state/form'
+import { Form } from '../models/form'
 
 test('computeState', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   expect(state.getState()).toEqual({
-    displayName: 'FormState',
+    displayName: 'Form',
     editable: undefined,
     valid: true,
     invalid: false,
@@ -89,7 +89,7 @@ test('computeState', () => {
   expect(state.getState().unmounted).toEqual(true)
 })
 test('subscribe/unsubscribe', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   const cb = jest.fn()
   const idx = state.subscribe(cb)
   const paylaod = state.getState()
@@ -102,7 +102,7 @@ test('subscribe/unsubscribe', () => {
   expect(cb).toBeCalledWith(paylaod)
 })
 test('batch', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   const cb = jest.fn()
   state.batch(cb)
   expect(cb).toBeCalledTimes(1)
@@ -110,7 +110,7 @@ test('batch', () => {
   // force run getState
   const susCb = jest.fn()
   state.subscribe(susCb)
-  state.dirtyNum = 1
+  state.dirtyCount = 1
   state.batch(cb)
   expect(cb).toBeCalledTimes(2)
   expect(cb).toBeCalledWith()
@@ -118,7 +118,7 @@ test('batch', () => {
   expect(susCb).toBeCalledWith(state.state)
 })
 test('getState', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   const cb = jest.fn()
   state.getState(cb)
   expect(cb).toBeCalledTimes(1)
@@ -126,13 +126,13 @@ test('getState', () => {
   const syncState = state.getState()
   expect(syncState).toEqual(state.state)
 
-  state.controller.publishState = () => null
+  state.factory.getState = () => null
   state.getState(cb)
   expect(cb).toBeCalledTimes(2)
   expect(cb).toBeCalledWith(null)
 })
 test('setState', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   const susCb = jest.fn()
   state.subscribe(susCb)
   const cb1 = draft => {
@@ -197,7 +197,7 @@ test('setState', () => {
   })
 })
 test('getSourceState', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   const cb = jest.fn()
   state.getSourceState(cb)
   expect(cb).toBeCalledTimes(1)
@@ -205,13 +205,13 @@ test('getSourceState', () => {
   const syncState = state.getSourceState()
   expect(syncState).toEqual(state.state)
 
-  state.controller.publishState = () => null
+  state.factory.getState = () => null
   state.getSourceState(cb)
   expect(cb).toBeCalledTimes(2)
   expect(cb).toBeCalledWith(state.state)
 })
 test('setSourceState', () => {
-  const state = new FormState({ useDirty: false })
+  const state = new Form()
   const cb1 = draft => (draft.change = true)
   const prevState1 = state.getSourceState()
   expect(prevState1.change).toEqual(undefined)
@@ -220,12 +220,12 @@ test('setSourceState', () => {
   expect(state.getSourceState()).toEqual({ ...prevState1, change: true })
 })
 test('isDirty', () => {
-  const state = new FormState({ useDirty: true })
-  expect(state.dirtyNum).toEqual(0)
+  const state = new Form()
+  expect(state.dirtyCount).toEqual(0)
   expect(state.isDirty()).toEqual(false)
-  state.dirtyNum = 1
+  state.dirtyCount = 1
   expect(state.isDirty()).toEqual(true)
-  state.dirtyNum = 0
+  state.dirtyCount = 0
   expect(state.isDirty()).toEqual(false)
   state.dirtys.validating = true
   expect(state.isDirty()).toEqual(false)
