+++ packages/core/src/__tests__/vfield.state.spec.ts
@@ -1,7 +1,7 @@
-import { VirtualFieldState } from '../state/virtual-field'
+import { VirtualField } from '../models/virtual-field'
 
 test('computeState', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   state.setState((draft) => {
     draft.unmounted = true
   })
@@ -26,7 +26,7 @@ test('computeState', () => {
 
 
 test('subscribe/unsubscribe', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   const cb = jest.fn()
   const idx = state.subscribe(cb)
   const paylaod = state.getState()
@@ -39,7 +39,7 @@ test('subscribe/unsubscribe', () => {
   expect(cb).toBeCalledWith(paylaod)
 })
 test('batch', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   const cb = jest.fn()
   state.batch(cb)
   expect(cb).toBeCalledTimes(1)
@@ -47,7 +47,7 @@ test('batch', () => {
   // force run getState
   const susCb = jest.fn()
   state.subscribe(susCb)
-  state.dirtyNum = 1
+  state.dirtyCount = 1
   state.batch(cb)
   expect(cb).toBeCalledTimes(2)
   expect(cb).toBeCalledWith()
@@ -55,7 +55,7 @@ test('batch', () => {
   expect(susCb).toBeCalledWith(state.state)
 })
 test('getState', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   const cb = jest.fn()
   state.getState(cb)
   expect(cb).toBeCalledTimes(1)
@@ -63,13 +63,13 @@ test('getState', () => {
   const syncState = state.getState()
   expect(syncState).toEqual(state.state)
 
-  state.controller.publishState = () => null
+  state.factory.getState = () => null
   state.getState(cb)
   expect(cb).toBeCalledTimes(2)
   expect(cb).toBeCalledWith(null)
 })
 test('setState', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   const susCb = jest.fn()
   state.subscribe(susCb)
   const cb1 = (draft) => { draft.visible = false }
@@ -111,7 +111,7 @@ test('setState', () => {
   expect(susCb).toBeCalledWith({ ...prevState3, initialized: true, mounted: true })
 })
 test('getSourceState', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   const cb = jest.fn()
   state.getSourceState(cb)
   expect(cb).toBeCalledTimes(1)
@@ -119,13 +119,13 @@ test('getSourceState', () => {
   const syncState = state.getSourceState()
   expect(syncState).toEqual(state.state)
 
-  state.controller.publishState = () => null
+  state.factory.getState = () => null
   state.getSourceState(cb)
   expect(cb).toBeCalledTimes(2)
   expect(cb).toBeCalledWith(state.state)
 })
 test('setSourceState', () => {
-  const state = new VirtualFieldState({ useDirty: false })
+  const state = new VirtualField()
   const cb1 = (draft) => draft.change = true
   const prevState1 = state.getSourceState()
   expect(prevState1.change).toEqual(undefined)
@@ -134,12 +134,12 @@ test('setSourceState', () => {
   expect(state.getSourceState()).toEqual({ ...prevState1, change: true })
 })
 test('isDirty', () => {
-  const state = new VirtualFieldState({ useDirty: true })
-  expect(state.dirtyNum).toEqual(0)
+  const state = new VirtualField()
+  expect(state.dirtyCount).toEqual(0)
   expect(state.isDirty()).toEqual(false)
-  state.dirtyNum = 1
+  state.dirtyCount = 1
   expect(state.isDirty()).toEqual(true)
-  state.dirtyNum = 0
+  state.dirtyCount = 0
   expect(state.isDirty()).toEqual(false)
   state.dirtys.visible = true
   expect(state.isDirty()).toEqual(false)
